Move FFIClient/FFIEvents/FFIHandle/NativeMethods/FfiInstruction into LiveKit.Internal.FFI
Final step of the Internal/ namespace alignment. These 5 files live in
Internal/FFI/ but still declared LiveKit.Internal (or, for FfiInstruction,
flat LiveKit) — the last folder/namespace mismatch. Now they match their
folder and their FFI sibling files.

- Declarations -> LiveKit.Internal.FFI.
- Added `using LiveKit.Internal.FFI;` to the consumers that resolved these
  types via `using LiveKit.Internal;` (or same-namespace for FfiInstruction):
  ~18 Runtime files and 8 test files.
- FfiRequestWrap.cs: updated a fully-qualified `LiveKit.Internal.FfiClient`.
- YieldInstruction.cs: imports LiveKit.Internal.FFI (uses FfiClient.Instance).
- FFIClient.cs: moving into LiveKit.Internal.FFI brought the child `Pools`
  namespace into scope, which shadowed the `Pools` *class* (Pools.Pools).
  Aliased it (FfiResponsePools) at the one call site.

No behavioural change. Verified by compiling LiveKit, LiveKit.UniTask,
EditModeTests, PlayModeTests and PlayModeTests.UniTask via csc.
