+++ Runtime/Scripts/Audio/PlatformAudioSource.cs
@@ -3,6 +3,7 @@ using LiveKit.Proto;
 using LiveKit.Internal;
 using LiveKit.Internal.FFI.Requests;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit
 {
     /// <summary>
