+++ Runtime/Scripts/Core/E2EE.cs
@@ -3,6 +3,7 @@ using LiveKit.Internal;
 using LiveKit.Internal.FFI.Requests;
 using LiveKit.Proto;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit
 {
     public enum EncryptionType
