+++ Runtime/Scripts/Core/TrackPublication.cs
@@ -2,6 +2,7 @@ using LiveKit.Internal;
 using LiveKit.Internal.FFI.Requests;
 using LiveKit.Proto;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit
 {
     public class TrackPublication
