+++ Runtime/Scripts/DataStreams/ByteDataStream.cs
@@ -5,6 +5,7 @@ using LiveKit.Internal.FFI.Requests;
 using LiveKit.Internal;
 using LiveKit.Proto;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit
 {
     /// <summary>
