+++ Runtime/Scripts/DataStreams/TextDataStream.cs
@@ -4,6 +4,7 @@ using LiveKit.Internal.FFI.Requests;
 using LiveKit.Internal;
 using LiveKit.Proto;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit
 {
     /// <summary>
