+++ Tests/EditMode/FfiHandleTests.cs
@@ -2,6 +2,7 @@ using System;
 using LiveKit.Internal;
 using NUnit.Framework;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit.EditModeTests
 {
     public class FfiHandleTests
