+++ Tests/EditMode/SkipDispatchTests.cs
@@ -5,6 +5,7 @@ using LiveKit.Internal;
 using LiveKit.Proto;
 using NUnit.Framework;
 
+using LiveKit.Internal.FFI;
 namespace LiveKit.EditModeTests
 {
     public class SkipDispatchTests
